+++ CHANGELOG.md
@@ -17,8 +17,9 @@ solid fixes, as we are paving the way for the upcoming release of `1.0.0`. Here 
 - Prevent `paper` object from polluting the global scope (#1544).
 - Make sure `Path#arcTo()` always passes through the provide through point
   (#1477).
-- Draw shadows on `Raster` images (#1437)
+- Draw shadows on `Raster` images (#1437).
 - Fix boolean operation edge case (#1506, #1513, #1515).
+- Fix single segment closed path flatten error (#1338, #1569).
 - Remove memory leak on gradient colors (#1499).
 - Support alpha channel in CSS colors (#1468, #1539, #1565).
 - Improve color CSS string parsing and documentation.